Added support for floats

+++ header.cpp
@@ -2,7 +2,7 @@
 #include <string>
 #include <iostream>
 using namespace std;
-int change(int a, char op, int b){
+int change(float a, char op, float b){
     if (op == '+'){
         return a+b;
     }
@@ -10,14 +10,14 @@ int change(int a, char op, int b){
         return a-b;
     }
 }
-int bind(string a, char opa, string b, char opb){
+float bind(string a, char opa, string b, char opb){
     // 3 params: size, position and what you want to insert for overload
     a.insert(0, 1, opa);
     b.insert(0,1,opb);
-    int f = stoi(a)+stoi(b);
+    float f = stof(a)+stof(b);
     return f;
 }
-char check(int f){
+char check(float f){
    char s;
    if (f>0){
        s='+';
